% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMEFM-package.R
\docType{package}
\name{MMEFM-package}
\alias{MMEFM}
\alias{MMEFM-package}
\title{Multilevel Main Effects Matrix Factor Model Tools}
\description{
Estimation, rank selection, global common-factor detection and screening, and simulation for the Multilevel Main Effects Matrix Factor Model, including global and local main effects and common matrix-factor components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kx-liu/MMEFM}
  \item Report bugs at \url{https://github.com/kx-liu/MMEFM/issues}
}

}
\author{
\strong{Maintainer}: Kaixin Liu \email{k.liu31@lse.ac.uk}

Authors:
\itemize{
  \item Kaixin Liu \email{k.liu31@lse.ac.uk}
}

}
\keyword{internal}
