% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-detection.R
\name{print.mmefm_global_detection}
\alias{print.mmefm_global_detection}
\title{Print global common-factor detection}
\usage{
\method{print}{mmefm_global_detection}(x, ...)
}
\arguments{
\item{x}{An \code{mmefm_global_detection} object.}

\item{...}{Further arguments, currently unused.}
}
\value{
The input invisibly.
}
\description{
Print global common-factor detection
}
